refactor: rename auth services
